feat(mobile): Evolution E11 — premium video oynatıcı
KÜTÜPHANE KARARI (DoD gereği, pub.dev verisiyle): video_player KORUNDU, denetim
katmanı elle yazıldı. chewie kendi görsel dilini getiriyor + belgelenmiş arabellek
hatası var; better_player_plus önbellek/PiP veriyor ama 168 beğeni ve "kırıcı
değişiklik" uyarısı taşıyor; media_kit libmpv gömüyor (APK büyür) ve istemediğimiz
depolama izinlerini istiyor.

Mimari:
· PlaybackController soyutlaması — VideoPlayerController platform kanalına bağlı ve
  widget testinde örneklenemiyor. Kod tabanının yerleşik deseni (arayüz + uygulama)
  oynatıcıya da uygulandı → BÜTÜN denetimler sahte oynatıcıyla test ediliyor.
· Saf mantık domain/video/: WebVTT çözümleyici, devam/izlendi kuralları, bölüm
  eşleme, yer imi, biçimlendirme — video_player'a hiç dokunmuyor.
· İlerleme/yer imi CİHAZDA kalır (SharedPreferences); sunucuya taşımak E8'in
  gizlilik yükünü gereksiz büyütürdü.

Yüzey: bölüm işaretli sürüklenebilir zaman çizgisi (arabellek + yer imi işaretleri),
altyazı aç/kapa, hız döngüsü, 10 sn atlama, tam ekran + yönelim, bölüm listesi,
yer imleri, kaldığın yerden devam, "izlendi" durumu.

Yol boyunca düzeltilen ÜÇ gerçek kusur:
1. Kesirli bölüm zaman damgası sessizce kırpılıyordu — içerik t: 2.7 taşıyor, model
   int ilan etmişti, toInt() 2.7'yi 2 yapıyordu (9 sn'lik videoda ~%8 hata).
   Model double yapıldı, atlama milisaniye üzerinden.
2. Altyazı kutusu satır içi oynatıcıda ortadaki düğmelerin üstüne biniyordu.
3. "Tam ekran" tam ekran değildi — shell gezginine itildiği için alt sekme çubuğu
   kalıyordu; rootNavigator: true ile düzeltildi.

Dürüstçe yapılmayanlar (raporda gerekçeli): PiP, çevrimdışı indirme,
parlaklık/ses hareketleri, wakelock.

flutter analyze 0 · flutter test 263 (+59) · lint/format temiz · APK 69,9 MiB
Cihazda 15 maddelik denetim doğrulaması yapıldı (AYXSUKIVJVPZ7HPZ).

+++ apps/mobile/lib/domain/content/video_content.dart
@@ -6,14 +6,16 @@ part 'video_content.g.dart';
 /// Video bölümü (zaman damgalı içindekiler) (web `VideoChapter`).
 @freezed
 abstract class VideoChapter with _$VideoChapter {
-  const factory VideoChapter({required int t, required String title}) = _VideoChapter;
+  /// Saniye — KESİRLİ olabilir (içerik `2.7` gibi değerler taşıyor).
+  const factory VideoChapter({required double t, required String title}) = _VideoChapter;
   factory VideoChapter.fromJson(Map<String, Object?> json) => _$VideoChapterFromJson(json);
 }
 
 /// Transkript satırı (web `TranscriptCue`).
 @freezed
 abstract class TranscriptCue with _$TranscriptCue {
-  const factory TranscriptCue({required int t, required String text}) = _TranscriptCue;
+  /// Saniye — KESİRLİ olabilir.
+  const factory TranscriptCue({required double t, required String text}) = _TranscriptCue;
   factory TranscriptCue.fromJson(Map<String, Object?> json) => _$TranscriptCueFromJson(json);
 }
